Move some startup logic from MainActivity to SplashFragment
