+++ simplified-ui-splash/build.gradle
@@ -7,6 +7,7 @@ dependencies {
   api project(":simplified-reports")
   api project(":simplified-ui-branding")
   api project(":simplified-ui-theme")
+  api project(":simplified-services-api")
 
   implementation libraries.androidx_appcompat
   implementation libraries.androidx_constraintlayout
